Extract fetchHistoryHandler logic into a new file
